refactor(dashboard): migrate Chat + Sessions + chat internals — app.js shrinks 1562 → 125 LoC
Stage 4 PR 4.6 from #28. The largest single migration: ChatPanel +
SessionsPanel + all 7 modals + ToolCard + ChatMessage + diff helpers
move out of app.js into typed modules.

- components/chat-internals.ts (541 LoC) — ROLE_GLYPH, ToolCard,
  ChatMessage, ModalCard, 7 modal components (ShellModal,
  ChoiceModal, PlanModal, EditReviewModal, WorkspaceModal,
  CheckpointModal, RevisionModal), plus lineDiff/pairDiffRows
  helpers used by EditReviewModal. Carries `// @ts-nocheck` for
  bulk migration; type tightening is a follow-up.
- panels/chat.ts (784 LoC) — ChatPanel + summarizeActiveTool +
  InFlightRow + ChatStatusBar. Same @ts-nocheck rationale.
- panels/sessions.ts (128 LoC) — SessionsPanel; depends on
  ChatMessage from chat-internals. Properly typed.

Bug fix: ROLE_GLYPH was accidentally dropped during PR 4.1 (small
panels) when the strip cut its definition along with the chat
banner comment. ChatMessage would have crashed at runtime with
ReferenceError on any chat / sessions render. Tests didn't catch
it because no test loads the bundled dashboard. Restored in
chat-internals.ts.

Cleanup: dashboard/app.js trimmed to a 125-line shell — imports +
TABS array + the App component (sidebar + tab routing + error
boundary + toast stack). The @ts-nocheck pragmas in chat-related
files are explicitly tagged "tighten in follow-up".

Bundle: 122 → 123 KB (ROLE_GLYPH restoration). 1682 tests pass.
