+++ dashboard/app.js
@@ -1,30 +1,12 @@
 // Reasonix dashboard SPA — Preact 10 + HTM, bundled by tsup. CDN imports stay external.
 
-import hljs from "https://esm.sh/highlight.js@11.10.0/lib/common";
 import htm from "https://esm.sh/htm@3.1.1";
 import { h, render } from "https://esm.sh/preact@10.22.0";
-import {
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from "https://esm.sh/preact@10.22.0/hooks";
-import { MetricCard } from "./src/components/metric-card";
-import { MODE, TOKEN, api } from "./src/lib/api";
-import { ToastStack, appBus, showToast, toastBus } from "./src/lib/bus";
+import { useCallback, useEffect, useState } from "https://esm.sh/preact@10.22.0/hooks";
+import { MODE } from "./src/lib/api";
+import { ToastStack, appBus } from "./src/lib/bus";
 import { ErrorBoundary, ErrorOverlay } from "./src/lib/error-boundary";
-import { fmtBytes, fmtNum, fmtPct, fmtRelativeTime, fmtUsd } from "./src/lib/format";
-import {
-  escapeHtml,
-  hlLine,
-  langFromPath,
-  renderHighlightedBlock,
-  renderMarkdownToString,
-  renderSearchReplace,
-  renderUnifiedDiff,
-} from "./src/lib/markdown";
-import { usePoll } from "./src/lib/use-poll";
+import { ChatPanel } from "./src/panels/chat";
 import { HooksPanel } from "./src/panels/hooks";
 import { McpPanel } from "./src/panels/mcp";
 import { MemoryPanel } from "./src/panels/memory";
@@ -32,6 +14,7 @@ import { OverviewPanel } from "./src/panels/overview";
 import { PermissionsPanel } from "./src/panels/permissions";
 import { PlansPanel } from "./src/panels/plans";
 import { SemanticPanel } from "./src/panels/semantic";
+import { SessionsPanel } from "./src/panels/sessions";
 import { SettingsPanel } from "./src/panels/settings";
 import { SkillsPanel } from "./src/panels/skills";
 import { SystemPanel } from "./src/panels/system";
@@ -40,1525 +23,27 @@ import { UsagePanel } from "./src/panels/usage";
 
 const html = htm.bind(h);
 
-// ---------- panels ----------
-
-function renderMessageBody(text) {
-  if (!text) return null;
-  return html`<div class="md" dangerouslySetInnerHTML=${{ __html: renderMarkdownToString(text) }}></div>`;
-}
-
-function parseToolArgs(raw) {
-  if (!raw) return null;
-  try {
-    return JSON.parse(raw);
-  } catch {
-    return null;
-  }
-}
-
-function ToolCard({ msg }) {
-  const args = parseToolArgs(msg.toolArgs);
-  const name = msg.toolName ?? "tool";
-  // Reasonix's filesystem tools emit the path in args.path; MCP-bridged
-  // ones may differ but most expose a `path` field too. Normalize.
-  const path = args?.path ?? args?.file_path ?? args?.filename;
-
-  // edit_file (Reasonix) — search/replace pair → diff view.
-  if (
-    (name === "edit_file" || name.endsWith("_edit_file")) &&
-    args &&
-    typeof args.search === "string" &&
-    typeof args.replace === "string"
-  ) {
-    const diffHtml = renderSearchReplace(args.search, args.replace, path ?? "");
-    return html`
-      <div class="tool-card">
-        <div class="tool-card-head">
-          <span class="tool-card-icon">✎</span>
-          <span class="tool-card-name">edit_file</span>
-          ${path ? html`<code class="tool-card-path">${path}</code>` : null}
-        </div>
-        <div dangerouslySetInnerHTML=${{ __html: diffHtml }}></div>
-        ${msg.text ? html`<div class="tool-card-result">${msg.text}</div>` : null}
-      </div>
-    `;
-  }
-
-  // write_file — show new content as a code block with path-derived lang.
-  if (
-    (name === "write_file" || name.endsWith("_write_file")) &&
-    args &&
-    typeof args.content === "string"
-  ) {
-    const lang = langFromPath(path);
-    return html`
-      <div class="tool-card">
-        <div class="tool-card-head">
-          <span class="tool-card-icon">+</span>
-          <span class="tool-card-name">write_file</span>
-          ${path ? html`<code class="tool-card-path">${path}</code>` : null}
-          ${lang ? html`<span class="pill pill-dim">${lang}</span>` : null}
-        </div>
-        <div dangerouslySetInnerHTML=${{ __html: renderHighlightedBlock(args.content, lang) }}></div>
-        ${msg.text ? html`<div class="tool-card-result">${msg.text}</div>` : null}
-      </div>
-    `;
-  }
-
-  // read_file / list_files — content lands in msg.text.
-  if (name === "read_file" || name.endsWith("_read_file") || name === "filesystem_read_file") {
-    const lang = langFromPath(path);
-    return html`
-      <div class="tool-card">
-        <div class="tool-card-head">
-          <span class="tool-card-icon">▤</span>
-          <span class="tool-card-name">read_file</span>
-          ${path ? html`<code class="tool-card-path">${path}</code>` : null}
-          ${lang ? html`<span class="pill pill-dim">${lang}</span>` : null}
-        </div>
-        <div dangerouslySetInnerHTML=${{ __html: renderHighlightedBlock(msg.text, lang) }}></div>
-      </div>
-    `;
-  }
-
-  // run_command / run_background — terminal-style.
-  if (name === "run_command" || name === "run_background") {
-    const cmd = args?.command;
-    return html`
-      <div class="tool-card">
-        <div class="tool-card-head">
-          <span class="tool-card-icon">⚡</span>
-          <span class="tool-card-name">${name === "run_background" ? "run_background" : "run_command"}</span>
-        </div>
-        ${
-          cmd
-            ? html`<pre class="tool-card-cmd"><span class="tool-card-prompt">$</span> <code>${cmd}</code></pre>`
-            : null
-        }
-        ${msg.text ? html`<pre class="tool-card-output">${msg.text}</pre>` : null}
-      </div>
-    `;
-  }
-
-  // list_files / file_exists / delete_file — show args + result inline.
-  if (
-    name === "list_files" ||
-    name === "file_exists" ||
-    name === "delete_file" ||
-    name === "create_directory" ||
-    name === "delete_directory" ||
-    name.endsWith("_list_files")
-  ) {
-    return html`
-      <div class="tool-card">
-        <div class="tool-card-head">
-          <span class="tool-card-icon">▣</span>
-          <span class="tool-card-name">${name}</span>
-          ${path ? html`<code class="tool-card-path">${path}</code>` : null}
-        </div>
-        <pre class="tool-card-output">${msg.text}</pre>
-      </div>
-    `;
-  }
-
-  // Default — keep the legacy compact box but add an args preview when
-  // present so MCP-bridged tools still surface something readable.
-  return html`
-    <div class="tool-card">
-      <div class="tool-card-head">
-        <span class="tool-card-icon">▣</span>
-        <span class="tool-card-name">${name}</span>
-      </div>
-      ${
-        args
-          ? html`<details class="tool-card-args"><summary>arguments</summary><pre>${escapeHtml(JSON.stringify(args, null, 2))}</pre></details>`
-          : null
-      }
-      <pre class="tool-card-output">${msg.text}</pre>
-    </div>
-  `;
-}
-
-function ChatMessage({ msg, streaming }) {
-  const role = msg.role;
-  const glyph = ROLE_GLYPH[role] ?? "·";
-  if (role === "tool") {
-    return html`
-      <div class="chat-msg tool">
-        <div class="glyph">${glyph}</div>
-        <${ToolCard} msg=${msg} />
-      </div>
-    `;
-  }
-  return html`
-    <div class="chat-msg ${role}">
-      <div class="glyph">${glyph}</div>
-      <div class="body">
-        ${msg.reasoning ? html`<div class="reasoning">${msg.reasoning}</div>` : null}
-        ${renderMessageBody(msg.text)}
-        ${streaming ? html`<span class="chat-streaming-cursor"></span>` : null}
-      </div>
-    </div>
-  `;
-}
-
-// ---------- Modal components mirroring the TUI ----------
-//
-// Each component renders a card matching the TUI's ModalCard accent
-// palette: red for shell (run-now), magenta for choice (branching),
-// cyan for plan (decision), green for edits. onResolve pushes to the
-// server; the SSE channel will echo back a modal-down that clears the
-// local state — both surfaces stay in lockstep without polling.
-
-function ModalCard({ accent, icon, title, subtitle, children }) {
-  return html`
-    <div class="modal-card" style=${`border-left-color: ${accent};`}>
-      <div class="modal-card-head">
-        <span class="modal-card-icon" style=${`color: ${accent};`}>${icon}</span>
-        <div>
-          <div class="modal-card-title">${title}</div>
-          ${subtitle ? html`<div class="modal-card-subtitle">${subtitle}</div>` : null}
-        </div>
-      </div>
-      ${children}
-    </div>
-  `;
-}
-
-function ShellModal({ modal, onResolve }) {
-  const isBg = modal.shellKind === "run_background";
-  return html`
-    <${ModalCard}
-      accent="#f87171"
-      icon=${isBg ? "⏱" : "⚡"}
-      title=${isBg ? "background process" : "shell command"}
-      subtitle=${
-        isBg ? "long-running — keeps running after approval" : "model wants to run a shell command"
-      }
-    >
-      <div class="modal-cmd"><span class="modal-cmd-prompt">$</span> <code>${modal.command}</code></div>
-      <div class="modal-actions">
-        <button class="primary" onClick=${() => onResolve("shell", "run_once")}>Run once</button>
-        <button onClick=${() => onResolve("shell", "always_allow")}>Always allow "${modal.allowPrefix}"</button>
-        <button class="danger" onClick=${() => onResolve("shell", "deny")}>Deny</button>
-      </div>
-    <//>
-  `;
-}
-
-function ChoiceModal({ modal, onResolve }) {
-  const [custom, setCustom] = useState("");
-  const [showCustom, setShowCustom] = useState(false);
-  return html`
-    <${ModalCard} accent="#f0abfc" icon="🔀" title="model wants you to pick" subtitle=${modal.question}>
-      ${modal.options.map(
-        (opt) => html`
-        <button
-          key=${opt.id}
-          class="modal-choice-row"
-          onClick=${() => onResolve("choice", { kind: "pick", optionId: opt.id })}
-        >
-          <span class="modal-choice-id">${opt.id}</span>
-          <span class="modal-choice-title">${opt.title}</span>
-          ${opt.summary ? html`<span class="modal-choice-summary">${opt.summary}</span>` : null}
-        </button>
-      `,
-      )}
-      ${
-        modal.allowCustom
-          ? showCustom
-            ? html`
-            <div class="modal-custom">
-              <textarea
-                placeholder="Type a free-form answer…"
-                rows="2"
-                value=${custom}
-                onInput=${(e) => setCustom(e.target.value)}
-              ></textarea>
-              <div class="modal-actions">
-                <button class="primary" onClick=${() => onResolve("choice", { kind: "custom", text: custom })} disabled=${!custom.trim()}>Send</button>
-                <button onClick=${() => {
-                  setShowCustom(false);
-                  setCustom("");
-                }}>Back</button>
-              </div>
-            </div>
-          `
-            : html`
-            <button class="modal-choice-row" onClick=${() => setShowCustom(true)}>
-              <span class="modal-choice-id">·</span>
-              <span class="modal-choice-title">Type my own answer</span>
-              <span class="modal-choice-summary">None of the above fits — write a free-form reply.</span>
-            </button>
-          `
-          : null
-      }
-      <button class="modal-choice-row modal-choice-cancel" onClick=${() => onResolve("choice", { kind: "cancel" })}>
-        <span class="modal-choice-id">×</span>
-        <span class="modal-choice-title">Cancel</span>
-        <span class="modal-choice-summary">Drop the question. Model will ask what you actually want.</span>
-      </button>
-    <//>
-  `;
-}
-
-function PlanModal({ modal, onResolve }) {
-  const [feedback, setFeedback] = useState("");
-  const [stage, setStage] = useState(null); // null | "approve" | "refine"
-  const send = () => onResolve("plan", stage, feedback);
-  return html`
-    <${ModalCard} accent="#67e8f9" icon="◆" title="plan submitted" subtitle="model proposed a plan; review then pick">
-      <div class="md modal-plan-body" dangerouslySetInnerHTML=${{ __html: marked.parse(modal.body || "") }}></div>
-      ${
-        stage
-          ? html`
-          <textarea
-            placeholder=${
-              stage === "approve"
-                ? "Optional last instructions / answers to open questions (Enter to send blank)"
-                : "What needs to change? Be specific."
-            }
-            rows="3"
-            value=${feedback}
-            onInput=${(e) => setFeedback(e.target.value)}
-          ></textarea>
-          <div class="modal-actions">
-            <button class="primary" onClick=${send}>${stage === "approve" ? "Approve" : "Send refinement"}</button>
-            <button onClick=${() => {
-              setStage(null);
-              setFeedback("");
-            }}>Back</button>
-          </div>
-        `
-          : html`
-          <div class="modal-actions">
-            <button class="primary" onClick=${() => setStage("approve")}>Approve</button>
-            <button onClick=${() => setStage("refine")}>Refine</button>
-            <button class="danger" onClick=${() => onResolve("plan", "cancel")}>Cancel</button>
-          </div>
-        `
-      }
-    <//>
-  `;
-}
-
-// Line-level LCS diff. Returns an ordered list of rows; "context" rows
-// appear on both sides, "del" only on the left (red), "ins" only on the
-// right (green). Adjacent del/ins are paired into one row downstream so
-// the change reads "old → new" left-to-right like a git side-by-side.
-function lineDiff(aLines, bLines) {
-  const m = aLines.length;
-  const n = bLines.length;
-  const dp = Array.from({ length: m + 1 }, () => new Array(n + 1).fill(0));
-  for (let i = 1; i <= m; i++) {
-    for (let j = 1; j <= n; j++) {
-      if (aLines[i - 1] === bLines[j - 1]) dp[i][j] = dp[i - 1][j - 1] + 1;
-      else dp[i][j] = Math.max(dp[i - 1][j], dp[i][j - 1]);
-    }
-  }
-  const out = [];
-  let i = m;
-  let j = n;
-  while (i > 0 || j > 0) {
-    if (i > 0 && j > 0 && aLines[i - 1] === bLines[j - 1]) {
-      out.push({ kind: "context", text: aLines[i - 1] });
-      i--;
-      j--;
-    } else if (j > 0 && (i === 0 || dp[i][j - 1] >= dp[i - 1][j])) {
-      out.push({ kind: "ins", text: bLines[j - 1] });
-      j--;
-    } else {
-      out.push({ kind: "del", text: aLines[i - 1] });
-      i--;
-    }
-  }
-  return out.reverse();
-}
-
-// Pair del/ins runs into side-by-side rows. A run of consecutive dels
-// followed by a run of inss collapses into rows of (del[k], ins[k]) so
-// the modified line lines up across the gutter; surplus on either side
-// produces rows with the opposite cell empty.
-function pairDiffRows(diff) {
-  const rows = [];
-  let k = 0;
-  while (k < diff.length) {
-    if (diff[k].kind === "context") {
-      rows.push({ left: diff[k].text, right: diff[k].text, kind: "context" });
-      k++;
-      continue;
-    }
-    const dels = [];
-    const inss = [];
-    while (k < diff.length && diff[k].kind === "del") {
-      dels.push(diff[k].text);
-      k++;
-    }
-    while (k < diff.length && diff[k].kind === "ins") {
-      inss.push(diff[k].text);
-      k++;
-    }
-    const pairs = Math.max(dels.length, inss.length);
-    for (let p = 0; p < pairs; p++) {
-      rows.push({
-        left: dels[p] ?? null,
-        right: inss[p] ?? null,
-        kind: dels[p] != null && inss[p] != null ? "change" : dels[p] != null ? "del" : "ins",
-      });
-    }
-  }
-  return rows;
-}
-
-function EditReviewModal({ modal, onResolve }) {
-  const search = modal.search ?? "";
-  const replace = modal.replace ?? "";
-  const lang = langFromPath(modal.path);
-  const aLines = search.split("\n");
-  const bLines = replace.split("\n");
-  const rows = pairDiffRows(lineDiff(aLines, bLines));
-
-  return html`
-    <${ModalCard}
-      accent="#86efac"
-      icon="◆"
-      title="edit pending review"
-      subtitle=${`${modal.path} · ${modal.remaining} of ${modal.total} blocks remaining`}
-    >
-      <div class="edit-diff-wrap">
-        <div class="edit-diff-head">
-          <div class="edit-diff-side edit-diff-side-old">
-            <span class="edit-diff-marker">−</span> before
-          </div>
-          <div class="edit-diff-side edit-diff-side-new">
-            <span class="edit-diff-marker">+</span> after
-          </div>
-        </div>
-        <div class="edit-diff-body">
-          ${rows.map(
-            (row, i) => html`
-            <div key=${i} class=${`edit-diff-row edit-diff-row-${row.kind}`}>
-              <div class="edit-diff-cell edit-diff-cell-old">
-                ${
-                  row.left != null
-                    ? html`<span
-                        class="edit-diff-line"
-                        dangerouslySetInnerHTML=${{ __html: hlLine(row.left, lang) || "&nbsp;" }}
-                      ></span>`
-                    : html`<span class="edit-diff-empty">&nbsp;</span>`
-                }
-              </div>
-              <div class="edit-diff-cell edit-diff-cell-new">
-                ${
-                  row.right != null
-                    ? html`<span
-                        class="edit-diff-line"
-                        dangerouslySetInnerHTML=${{ __html: hlLine(row.right, lang) || "&nbsp;" }}
-                      ></span>`
-                    : html`<span class="edit-diff-empty">&nbsp;</span>`
-                }
-              </div>
-            </div>
-          `,
-          )}
-        </div>
-      </div>
-      <div class="modal-actions">
-        <button class="primary" onClick=${() => onResolve("edit-review", "apply")}>Apply (y)</button>
-        <button onClick=${() => onResolve("edit-review", "reject")}>Reject (n)</button>
-        <button onClick=${() => onResolve("edit-review", "apply-rest-of-turn")}>Apply rest (a)</button>
-        <button onClick=${() => onResolve("edit-review", "flip-to-auto")}>Flip to AUTO (A)</button>
-      </div>
-    <//>
-  `;
-}
-
-function WorkspaceModal({ modal, onResolve }) {
-  return html`
-    <${ModalCard}
-      accent="#fbbf24"
-      icon="◇"
-      title="model wants to switch workspace"
-      subtitle="every subsequent file / shell / memory tool resolves against the new root"
-    >
-      <div class="modal-cmd"><span class="modal-cmd-prompt">→</span> <code>${modal.path}</code></div>
-      <div class="modal-actions">
-        <button class="primary" onClick=${() => onResolve("workspace", "switch")}>Switch (Enter)</button>
-        <button class="danger" onClick=${() => onResolve("workspace", "deny")}>Deny (Esc)</button>
-      </div>
-    <//>
-  `;
-}
-
-function CheckpointModal({ modal, onResolve }) {
-  const [reviseText, setReviseText] = useState("");
-  const [staged, setStaged] = useState(false);
-  const label = modal.title ? `${modal.stepId} · ${modal.title}` : modal.stepId;
-  const counter = modal.total > 0 ? ` (${modal.completed}/${modal.total})` : "";
-  return html`
-    <${ModalCard}
-      accent="#a5f3fc"
-      icon="✓"
-      title=${`step complete${counter}`}
-      subtitle=${label}
-    >
-      ${
-        staged
-          ? html`
-          <textarea
-            placeholder="What needs to change before the next step? Leave blank to just continue."
-            rows="3"
-            value=${reviseText}
-            onInput=${(e) => setReviseText(e.target.value)}
-          ></textarea>
-          <div class="modal-actions">
-            <button class="primary" onClick=${() => onResolve("checkpoint", "revise", reviseText)}>Send revision</button>
-            <button onClick=${() => {
-              setStaged(false);
-              setReviseText("");
-            }}>Back</button>
-          </div>
-        `
-          : html`
-          <div class="modal-actions">
-            <button class="primary" onClick=${() => onResolve("checkpoint", "continue")}>Continue</button>
-            <button onClick=${() => setStaged(true)}>Revise…</button>
-            <button class="danger" onClick=${() => onResolve("checkpoint", "stop")}>Stop</button>
-          </div>
-        `
-      }
-    <//>
-  `;
-}
-
-function RevisionModal({ modal, onResolve }) {
-  const riskColor = (r) =>
-    r === "high" ? "#f87171" : r === "med" ? "#fbbf24" : r === "low" ? "#86efac" : "#9ca3af";
-  return html`
-    <${ModalCard}
-      accent="#c4b5fd"
-      icon="✎"
-      title="model proposed a plan revision"
-      subtitle=${modal.summary || modal.reason}
-    >
-      <div class="modal-revise-reason">${modal.reason}</div>
-      <ol class="modal-revise-steps">
-        ${modal.remainingSteps.map(
-          (s) => html`
-            <li key=${s.id}>
-              <span class="modal-revise-dot" style=${`background:${riskColor(s.risk)}`}></span>
-              <span class="modal-revise-id">${s.id}</span>
-              <span class="modal-revise-title">${s.title}</span>
-              <span class="modal-revise-action">${s.action}</span>
-            </li>
-          `,
-        )}
-      </ol>
-      <div class="modal-actions">
-        <button class="primary" onClick=${() => onResolve("revision", "accept")}>Accept</button>
-        <button class="danger" onClick=${() => onResolve("revision", "reject")}>Reject</button>
-      </div>
-    <//>
-  `;
-}
-
-function ChatPanel() {
-  const [messages, setMessages] = useState([]);
-  const [streaming, setStreaming] = useState(null); // { id, text, reasoning }
-  // Tool currently dispatched but not yet returning. Set on `tool_start`,
-  // cleared on `tool` / `error`. Drives the in-flight row so the user
-  // sees what's running (path, command, char counts) instead of a
-  // generic "waiting" — file writes especially feel hung otherwise.
-  const [activeTool, setActiveTool] = useState(null); // { id, toolName, args }
-  const [busy, setBusy] = useState(false);
-  const [input, setInput] = useState("");
-  const [error, setError] = useState(null);
-  const [bootError, setBootError] = useState(null);
-  const [statusLine, setStatusLine] = useState(null);
-  // Mirror of the active TUI modal: { kind, ...payload } | null. Set
-  // by `modal-up` SSE events, cleared by `modal-down`. Web uses POST
-  // /api/modal/resolve to drive resolution; either surface clears the
-  // other's modal via the resulting modal-down event.
-  const [modal, setModal] = useState(null);
-  // Current edit gate (review / auto / yolo). null when not in code
-  // mode. Refreshed via /api/overview poll because the mode also
-  // flips from TUI Shift+Tab and we want the segmented control to
-  // stay in sync without a dedicated event.
-  const [editMode, setEditModeLocal] = useState(null);
-  // Persisted preset + reasoning_effort, surfaced here so the user
-  // can flip them mid-chat without leaving the tab. /api/overview
-  // includes both since 0.14.x; the same poll covers all three.
-  const [preset, setPresetLocal] = useState(null);
-  const [effort, setEffortLocal] = useState(null);
-  // Live session stats — cache hit, costs, tokens, balance — from the
-  // same /api/overview poll. Renders into a compact status bar below
-  // the input area.
-  const [stats, setStats] = useState(null);
-  const [overviewModel, setOverviewModel] = useState(null);
-  // Whether the project has a built semantic index. Null = unknown
-  // (poll hasn't landed) or non-attached. False = no index → show the
-  // dismissible banner. True = index built → hide it.
-  const [semanticIndex, setSemanticIndex] = useState(null);
-  const [semanticBannerDismissed, setSemanticBannerDismissed] = useState(() => {
-    try {
-      return localStorage.getItem("rx.semanticBannerDismissed") === "1";
-    } catch {
-      return false;
-    }
-  });
-  useEffect(() => {
-    try {
-      localStorage.setItem("rx.semanticBannerDismissed", semanticBannerDismissed ? "1" : "0");
-    } catch {
-      /* ignore */
-    }
-  }, [semanticBannerDismissed]);
-  // Wall-clock timestamp the current turn started at — populated when
-  // busy flips true, cleared when it flips false. Drives the "elapsed
-  // Ns" readout in the in-flight indicator. Refreshed once per second
-  // by `nowTick` so the seconds counter ticks visibly even between
-  // SSE deltas.
-  const [turnStartedAt, setTurnStartedAt] = useState(null);
-  const [nowTick, setNowTick] = useState(0);
-  useEffect(() => {
-    if (!busy) return;
-    const id = setInterval(() => setNowTick((n) => n + 1), 500);
-    return () => clearInterval(id);
-  }, [busy]);
-  useEffect(() => {
-    if (busy) {
-      if (!turnStartedAt) setTurnStartedAt(Date.now());
-    } else {
-      setTurnStartedAt(null);
-    }
-  }, [busy, turnStartedAt]);
-  // Sticks to bottom only while the user is already near the bottom.
-  // Once they scroll up to read older content the streaming deltas no
-  // longer yank the view back. Re-armed when they scroll back to the
-  // bottom on their own. 80px threshold absorbs sub-pixel rounding.
-  const shouldAutoScroll = useRef(true);
-  // Ref to the scrollable feed container so we don't have to rely on
-  // a global querySelector (which would race the conditional render
-  // — `.chat-feed` only mounts when at least one message is present).
-  // The feed is now always rendered, so `feedRef.current` is set on
-  // first paint and the scroll listener attaches once.
-  const feedRef = useRef(null);
-
-  // Initial snapshot — messages + busy + any modal already up.
-  useEffect(() => {
-    let cancelled = false;
-    (async () => {
-      try {
-        const data = await api("/messages");
-        if (cancelled) return;
-        setMessages(data.messages ?? []);
-        setBusy(Boolean(data.busy));
-      } catch (err) {
-        if (!cancelled) setBootError(err.message);
-      }
-      try {
-        const m = await api("/modal");
-        if (!cancelled && m.modal) setModal(m.modal);
-      } catch {
-        /* skip — modal endpoint optional in standalone */
-      }
-    })();
-    return () => {
-      cancelled = true;
-    };
-  }, []);
-
-  // Live event stream.
-  useEffect(() => {
-    const es = new EventSource(`/api/events?token=${TOKEN}`);
-    es.onmessage = (ev) => {
-      let dash;
-      try {
-        dash = JSON.parse(ev.data);
-      } catch {
-        return;
-      }
-      if (dash.kind === "ping") return;
-      if (dash.kind === "busy-change") {
-        setBusy(dash.busy);
-        return;
-      }
-      if (dash.kind === "user") {
-        setMessages((prev) => [...prev, { id: dash.id, role: "user", text: dash.text }]);
-        return;
-      }
-      if (dash.kind === "assistant_delta") {
-        setStreaming((cur) => {
-          const text = (cur?.text ?? "") + (dash.contentDelta ?? "");
-          const reasoning = (cur?.reasoning ?? "") + (dash.reasoningDelta ?? "");
-          return { id: dash.id, text, reasoning };
-        });
-        return;
-      }
-      if (dash.kind === "assistant_final") {
-        setStreaming(null);
-        setMessages((prev) => [
-          ...prev,
-          {
-            id: dash.id,
-            role: "assistant",
-            text: dash.text,
-            reasoning: dash.reasoning,
-          },
-        ]);
-        return;
-      }
-      if (dash.kind === "tool_start") {
-        // Surface the dispatched tool + its args in the in-flight row.
-        // No info-row placeholder: the InFlightRow now renders the
-        // detail (path / command / char count) and the result card
-        // appears when the `tool` event lands. Two rows for one tool
-        // call was redundant noise.
-        setActiveTool({ id: dash.id, toolName: dash.toolName, args: dash.args });
-        return;
-      }
-      if (dash.kind === "tool") {
-        setActiveTool((cur) => (cur && cur.id === dash.id ? null : cur));
-        setMessages((prev) => [
-          ...prev,
-          {
-            id: dash.id,
-            role: "tool",
-            text: dash.content,
-            toolName: dash.toolName,
-            toolArgs: dash.args,
-          },
-        ]);
-        return;
-      }
-      if (dash.kind === "warning" || dash.kind === "error" || dash.kind === "info") {
-        if (dash.kind === "error") {
-          setActiveTool(null);
-        }
-        setMessages((prev) => [...prev, { id: dash.id, role: dash.kind, text: dash.text }]);
-        return;
-      }
-      if (dash.kind === "status") {
-        setStatusLine(dash.text);
-        // Clear the status line shortly so old hints don't pile up.
-        setTimeout(() => setStatusLine((cur) => (cur === dash.text ? null : cur)), 5000);
-        return;
-      }
-      if (dash.kind === "modal-up") {
-        setModal(dash.modal);
-        return;
-      }
-      if (dash.kind === "modal-down") {
-        setModal((cur) => (cur && cur.kind === dash.modalKind ? null : cur));
-        return;
-      }
-    };
-    es.onerror = () => {
-      // Auto-reconnect by default; surface a brief banner on persistent
-      // failure but don't tear down — EventSource retries in the background.
-      setError("event stream interrupted — reconnecting…");
-      setTimeout(() => setError(null), 3000);
-    };
-    return () => es.close();
-  }, []);
-
-  const send = useCallback(async () => {
-    const text = input.trim();
-    if (!text || busy) return;
-    setError(null);
-    try {
-      const res = await api("/submit", { method: "POST", body: { prompt: text } });
-      if (!res.accepted) {
-        setError(res.reason ?? "rejected");
-        return;
-      }
-      setInput("");
-    } catch (err) {
-      setError(err.message);
-    }
-  }, [input, busy]);
-
-  const abort = useCallback(async () => {
-    try {
-      await api("/abort", { method: "POST" });
-    } catch (err) {
-      setError(err.message);
-    }
-  }, []);
-
-  // /new wipes context + scrollback (server-side); /clear keeps the
-  // log but blanks the visible scroll. Both route through /api/submit
-  // because handleSubmit on the TUI side already parses slashes — keeps
-  // one source of truth, no special endpoint needed. Local messages
-  // state is reset optimistically; an /api/messages refetch reconciles.
-  const newConversation = useCallback(async () => {
-    if (busy) {
-      if (!confirm("A turn is in flight. Abort and start a new conversation?")) return;
-    } else if (messages.length > 0 && !confirm("Clear current conversation and start fresh?")) {
-      return;
-    }
-    try {
-      await api("/submit", { method: "POST", body: { prompt: "/new" } });
-      setMessages([]);
-      setStreaming(null);
-      setActiveTool(null);
-      showToast("new conversation", "info");
-      // Refetch to reconcile in case the slash queued an info row.
-      setTimeout(async () => {
-        try {
-          const r = await api("/messages");
-          setMessages(r.messages ?? []);
-        } catch {
-          /* swallow */
-        }
-      }, 200);
-    } catch (err) {
-      setError(`/new failed: ${err.message}`);
-    }
-  }, [busy, messages.length]);
-
-  const clearScrollback = useCallback(async () => {
-    try {
-      await api("/submit", { method: "POST", body: { prompt: "/clear" } });
-      setMessages([]);
-      setStreaming(null);
-      setActiveTool(null);
-      showToast("scrollback cleared", "info");
-      setTimeout(async () => {
-        try {
-          const r = await api("/messages");
-          setMessages(r.messages ?? []);
-        } catch {
-          /* swallow */
-        }
-      }, 200);
-    } catch (err) {
-      setError(`/clear failed: ${err.message}`);
-    }
-  }, []);
-
-  const onKeyDown = useCallback(
-    (e) => {
-      // Enter sends, Shift+Enter inserts newline.
-      if (e.key === "Enter" && !e.shiftKey) {
-        e.preventDefault();
-        send();
-      }
-    },
-    [send],
-  );
-
-  if (bootError) {
-    return html`<div class="notice err">chat unavailable: ${bootError}</div>`;
-  }
-
-  // Track whether the user is parked at the bottom. Update on every
-  // scroll event so a single wheel-up flips the auto-scroll guard
-  // immediately. The threshold is generous enough that overshoot
-  // (smooth-scroll rebound, sub-pixel rounding) doesn't accidentally
-  // re-arm tracking when the user is barely above bottom.
-  //
-  // We also distinguish *user* scroll events from auto-scroll's own
-  // programmatic `scrollTop = scrollHeight` writes. Without that gate
-  // the auto-scroll effect would briefly snap to bottom, fire its
-  // own scroll event, re-set shouldAutoScroll = true, then wonder
-  // why the user complained that they couldn't scroll up — because
-  // every wheel-up was racing against the next delta's auto-snap.
-  // We mark the ref as `auto-scrolling` for one tick around the
-  // programmatic write; the listener ignores events it sees during
-  // that window.
-  const autoScrollInFlight = useRef(false);
-  useEffect(() => {
-    const el = feedRef.current;
-    if (!el) return;
-    const onScroll = () => {
-      if (autoScrollInFlight.current) return;
-      const distFromBottom = el.scrollHeight - el.scrollTop - el.clientHeight;
-      shouldAutoScroll.current = distFromBottom < 80;
-    };
-    el.addEventListener("scroll", onScroll, { passive: true });
-    return () => el.removeEventListener("scroll", onScroll);
-  }, []);
-
-  // Auto-scroll only when the user hasn't scrolled away. Streaming
-  // deltas no longer yank the view back; manual wheel/drag wins.
-  useEffect(() => {
-    if (!shouldAutoScroll.current) return;
-    const el = feedRef.current;
-    if (!el) return;
-    autoScrollInFlight.current = true;
-    el.scrollTop = el.scrollHeight;
-    // Clear the gate after the browser has had a chance to fire the
-    // resulting scroll event (microtask-ish — rAF is overkill, a 0ms
-    // setTimeout is enough to land after the synchronous handler).
-    setTimeout(() => {
-      autoScrollInFlight.current = false;
-    }, 0);
-  }, [messages, streaming]);
-
-  const allMessages = streaming
-    ? [
-        ...messages,
-        {
-          id: streaming.id,
-          role: "assistant",
-          text: streaming.text,
-          reasoning: streaming.reasoning,
-        },
-      ]
-    : messages;
-
-  // Resolve the active modal via POST /api/modal/resolve. The server
-  // hands the choice straight to App.tsx's resolveXxx callback, which
-  // calls the same handler the TUI button would. The local `modal`
-  // state clears the moment the SSE channel echoes `modal-down`.
-  const resolveModal = useCallback(async (kind, choice, text) => {
-    try {
-      await api("/modal/resolve", {
-        method: "POST",
-        body: text !== undefined ? { kind, choice, text } : { kind, choice },
-      });
-    } catch (err) {
-      setError(`modal resolve failed: ${err.message}`);
-    }
-  }, []);
-
-  // Poll /api/overview for current edit mode. Polling (not SSE) is
-  // fine — the gate flips from /mode, Shift+Tab, AND the web button;
-  // a 4s poll is good enough to keep the segmented control visually
-  // honest without piping yet another event kind.
-  useEffect(() => {
-    let cancelled = false;
-    const tick = async () => {
-      try {
-        const o = await api("/overview");
-        if (cancelled) return;
-        setEditModeLocal(o.editMode ?? null);
-        setPresetLocal(o.preset ?? null);
-        setEffortLocal(o.reasoningEffort ?? null);
-        setStats(o.stats ?? null);
-        setOverviewModel(o.model ?? null);
-        setSemanticIndex(o.semanticIndexExists);
-      } catch {
-        /* swallow */
-      }
-    };
-    tick();
-    const t = setInterval(tick, 2500);
-    return () => {
-      cancelled = true;
-      clearInterval(t);
-    };
-  }, []);
-
-  const setEditMode = useCallback(async (next) => {
-    setEditModeLocal(next); // optimistic
-    try {
-      await api("/edit-mode", { method: "POST", body: { mode: next } });
-    } catch (err) {
-      setError(`mode switch failed: ${err.message}`);
-      try {
-        const o = await api("/overview");
-        setEditModeLocal(o.editMode ?? null);
-      } catch {
-        /* swallow */
-      }
-    }
-  }, []);
-
-  // Generic settings flipper for preset + effort. Both go through
-  // /api/settings (which writes to ~/.reasonix/config.json). preset
-  // applies next session, effort applies next turn — the buttons'
-  // tooltips remind the user.
-  const setSetting = useCallback(async (key, value) => {
-    if (key === "preset") setPresetLocal(value);
-    if (key === "reasoningEffort") setEffortLocal(value);
-    try {
-      await api("/settings", { method: "POST", body: { [key]: value } });
-    } catch (err) {
-      setError(`${key} switch failed: ${err.message}`);
-      try {
-        const o = await api("/overview");
-        setPresetLocal(o.preset ?? null);
-        setEffortLocal(o.reasoningEffort ?? null);
-      } catch {
-        /* swallow */
-      }
-    }
-  }, []);
-
-  return html`
-    <div class="chat-shell">
-      <div class="panel-header" style="margin-bottom: 12px;">
-        <h2 class="panel-title">Chat</h2>
-        <span class="panel-subtitle">
-          mirrors the live ${MODE === "attached" ? "TUI" : "session"} — type here or in the terminal, both surfaces stay in sync
-        </span>
-        <div class="header-pickers" style="margin-left: auto;">
-          ${
-            effort
-              ? html`
-              <div class="mode-picker" title="reasoning_effort — applies next turn">
-                ${["high", "max"].map(
-                  (e) => html`
-                  <button
-                    key=${e}
-                    class="mode-btn ${effort === e ? "active accent" : ""}"
-                    onClick=${() => setSetting("reasoningEffort", e)}
-                    title=${e === "max" ? "max (default — best quality)" : "high (cheaper / faster)"}
-                  >${e}</button>
-                `,
-                )}
-              </div>
-            `
-              : null
-          }
-          ${
-            preset
-              ? html`
-              <div class="mode-picker" title="preset — model commitment">
-                ${(() => {
-                  // Anything that isn't one of the three new presets
-                  // (including legacy fast/smart/max from old configs)
-                  // highlights as `auto` — the safe default. User can
-                  // re-pick explicitly if they want flash or pro.
-                  const KNOWN = ["auto", "flash", "pro"];
-                  const canonical = KNOWN.includes(preset) ? preset : "auto";
-                  return ["auto", "flash", "pro"].map(
-                    (p) => html`
-                      <button
-                        key=${p}
-                        class="mode-btn ${canonical === p ? "active accent" : ""}"
-                        onClick=${() => setSetting("preset", p)}
-                        title=${
-                          p === "auto"
-                            ? "auto — flash baseline; auto-escalates to pro on hard turns (NEEDS_PRO / failure threshold)"
-                            : p === "flash"
-                              ? "flash — always flash; no auto-escalate. /pro still works for one-shot manual"
-                              : "pro — always pro; ~3× flash cost (5/31 discount). Locks in on hard architecture work."
-                        }
-                      >${p}</button>
-                    `,
-                  );
-                })()}
-              </div>
-            `
-              : null
-          }
-          ${
-            editMode
-              ? html`
-              <div class="mode-picker" title="edit gate — Shift+Tab cycles in TUI">
-                ${["review", "auto", "yolo"].map(
-                  (m) => html`
-                  <button
-                    key=${m}
-                    class="mode-btn ${editMode === m ? "active" : ""} ${m === "yolo" ? "yolo" : ""}"
-                    onClick=${() => setEditMode(m)}
-                    title=${
-                      m === "review"
-                        ? "review — both edits and non-allowlisted shell ask first"
-                        : m === "auto"
-                          ? "auto — edits auto-apply, shell still asks"
-                          : "yolo — edits AND shell auto-run, allowlist bypassed"
-                    }
-                  >${m}</button>
-                `,
-                )}
-              </div>
-            `
-              : null
-          }
-        </div>
-      </div>
-
-      ${
-        !busy && statusLine
-          ? html`<div class="chat-status"><span class="muted">${statusLine}</span></div>`
-          : null
-      }
-      ${
-        semanticIndex === false && !semanticBannerDismissed
-          ? html`<div class="chat-banner">
-              <span class="chat-banner-icon">≈</span>
-              <span class="chat-banner-text">
-                <strong>Semantic search isn't enabled for this project.</strong>
-                <span class="muted">
-                  Build the index once and the model can find code by meaning ("where do we handle auth failures?") instead of grep on exact strings.
-                </span>
-              </span>
-              <button
-                class="primary"
-                onClick=${() => appBus.dispatchEvent(new CustomEvent("navigate-tab", { detail: { tabId: "semantic" } }))}
-              >Build it →</button>
-              <button
-                class="chat-banner-close"
-                onClick=${() => setSemanticBannerDismissed(true)}
-                title="dismiss (don't show again)"
-              >×</button>
-            </div>`
-          : null
-      }
-      ${error ? html`<div class="notice err">${error}</div>` : null}
-
-      ${
-        modal
-          ? modal.kind === "shell"
-            ? html`<${ShellModal} modal=${modal} onResolve=${resolveModal} />`
-            : modal.kind === "choice"
-              ? html`<${ChoiceModal} modal=${modal} onResolve=${resolveModal} />`
-              : modal.kind === "plan"
-                ? html`<${PlanModal} modal=${modal} onResolve=${resolveModal} />`
-                : modal.kind === "edit-review"
-                  ? html`<${EditReviewModal} modal=${modal} onResolve=${resolveModal} />`
-                  : modal.kind === "workspace"
-                    ? html`<${WorkspaceModal} modal=${modal} onResolve=${resolveModal} />`
-                    : modal.kind === "checkpoint"
-                      ? html`<${CheckpointModal} modal=${modal} onResolve=${resolveModal} />`
-                      : modal.kind === "revision"
-                        ? html`<${RevisionModal} modal=${modal} onResolve=${resolveModal} />`
-                        : null
-          : null
-      }
-
-      <div class="chat-feed" ref=${feedRef}>
-        ${
-          allMessages.length === 0
-            ? html`<div class="chat-empty">
-                No conversation yet. Send a prompt below to begin.
-              </div>`
-            : allMessages.map(
-                (m) => html`
-                  <${ChatMessage}
-                    key=${m.id}
-                    msg=${m}
-                    streaming=${streaming && streaming.id === m.id}
-                  />
-                `,
-              )
-        }
-      </div>
-
-      <div class="chat-input-area">
-        <textarea
-          placeholder=${busy ? "wait for the current turn to finish…" : "Type a prompt — Enter sends, Shift+Enter for a newline"}
-          value=${input}
-          onInput=${(e) => setInput(e.target.value)}
-          onKeyDown=${onKeyDown}
-          disabled=${busy}
-          rows="2"
-        ></textarea>
-        <div style="display: flex; flex-direction: column; gap: 6px; align-self: stretch; justify-content: flex-end;">
-          <button
-            class="primary"
-            onClick=${send}
-            disabled=${busy || !input.trim()}
-          >Send</button>
-          <div style="display: flex; gap: 6px;">
-            <button onClick=${newConversation} title="/new — wipe conversation context (loop log + scrollback)">New</button>
-            <button onClick=${clearScrollback} title="/clear — wipe just visible scrollback (context kept)">Clear</button>
-          </div>
-        </div>
-      </div>
-
-      ${
-        busy
-          ? html`<${InFlightRow}
-              streaming=${streaming}
-              activeTool=${activeTool}
-              startedAt=${turnStartedAt}
-              statusLine=${statusLine}
-              onAbort=${abort}
-              tick=${nowTick}
-            />`
-          : null
-      }
-      <${ChatStatusBar} stats=${stats} model=${overviewModel} />
-    </div>
-  `;
-}
-
-// Summarize the dispatched tool in one line — what the user wants to
-// know is "is this hung or really doing X". Per-tool projection so a
-// write_file says "→ /path/foo (12,345 ch)" instead of just "tool is
-// running". Returns null for tools we don't have a custom shape for;
-// the row falls back to the bare tool name.
-function summarizeActiveTool(activeTool) {
-  if (!activeTool) return null;
-  const name = activeTool.toolName ?? "tool";
-  const args = parseToolArgs(activeTool.args);
-  const path = args?.path ?? args?.file_path ?? args?.filename;
-  if (name === "write_file" && path) {
-    const len = typeof args?.content === "string" ? args.content.length : null;
-    return `${name} → ${path}${len != null ? ` (${len.toLocaleString()} ch)` : ""}`;
-  }
-  if ((name === "edit_file" || name.endsWith("_edit_file")) && path) {
-    return `${name} → ${path}`;
-  }
-  if ((name === "run_command" || name === "run_background") && typeof args?.command === "string") {
-    const c = args.command;
-    return `${name} → $ ${c.length > 80 ? `${c.slice(0, 80)}…` : c}`;
-  }
-  if ((name === "read_file" || name === "list_files" || name === "search_files") && path) {
-    return `${name} → ${path}`;
-  }
-  if (path) return `${name} → ${path}`;
-  return name;
-}
-
-// Live "what's the model doing right now" strip. Lives just above the
-// ChatStatusBar so the user's eyes don't have to leave the input area
-// to see whether the turn is alive — ticks every 500ms via the parent's
-// nowTick so the seconds counter shows visible motion even when the
-// SSE stream is silent (model thinking, waiting on a tool, etc).
-function InFlightRow({ streaming, activeTool, startedAt, statusLine, onAbort, tick: _tick }) {
-  const elapsedMs = startedAt ? Date.now() - startedAt : 0;
-  const elapsed = (elapsedMs / 1000).toFixed(1);
-  const reasoningLen = streaming?.reasoning?.length ?? 0;
-  const textLen = streaming?.text?.length ?? 0;
-  // Tool-running phase wins over text/reasoning since the model is
-  // blocked on the tool — even if assistant_delta has fired we want
-  // to show the active dispatch.
-  const toolSummary = summarizeActiveTool(activeTool);
-  const phase = toolSummary
-    ? "running"
-    : reasoningLen > 0 && textLen === 0
-      ? "thinking"
-      : textLen > 0
-        ? "streaming"
-        : "waiting";
-  return html`
-    <div class="chat-inflight">
-      <span class="spinner"></span>
-      <span class="chat-inflight-phase">${phase}</span>
-      <span class="chat-inflight-sep">·</span>
-      <span class="muted">${elapsed}s</span>
-      ${
-        toolSummary
-          ? html`
-            <span class="chat-inflight-sep">·</span>
-            <span class="chat-inflight-tool" title=${toolSummary}>${toolSummary}</span>
-          `
-          : null
-      }
-      ${
-        !toolSummary && (textLen > 0 || reasoningLen > 0)
-          ? html`
-            <span class="chat-inflight-sep">·</span>
-            <span class="muted">
-              ${reasoningLen > 0 ? html`reasoning ${reasoningLen.toLocaleString()} ch` : null}
-              ${reasoningLen > 0 && textLen > 0 ? html`<span> · </span>` : null}
-              ${textLen > 0 ? html`out ${textLen.toLocaleString()} ch` : null}
-            </span>
-          `
-          : null
-      }
-      ${
-        statusLine
-          ? html`
-            <span class="chat-inflight-sep">·</span>
-            <span class="muted">${statusLine}</span>
-          `
-          : null
-      }
-      <button class="chat-inflight-abort" onClick=${onAbort}>Abort (Esc)</button>
-    </div>
-  `;
-}
-
-// ---------- Chat status bar ----------
-//
-// Mirrors the TUI's StatsPanel — turn / session cost, cache hit %,
-// ctx token gauge, balance. Sits beneath the input area as a compact
-// monospace strip. Renders as a placeholder ("· · ·") while stats
-// haven't arrived yet so the layout doesn't shift on first paint.
-
-function ChatStatusBar({ stats, model }) {
-  if (!stats) {
-    return html`
-      <div class="chat-statusbar">
-        <span class="muted">· · ·  waiting for live stats</span>
-      </div>
-    `;
-  }
-  const ctxPct =
-    stats.contextCapTokens > 0 ? (stats.lastPromptTokens / stats.contextCapTokens) * 100 : 0;
-  const balance = stats.balance && stats.balance.length > 0 ? stats.balance[0] : null;
-  return html`
-    <div class="chat-statusbar">
-      <span class="status-item">
-        <span class="status-label">model</span>
-        <code>${model ?? "—"}</code>
-      </span>
-      <span class="status-item">
-        <span class="status-label">ctx</span>
-        <span class="status-bar-mini">
-          <span class="status-bar-mini-fill" style=${`width: ${Math.min(100, ctxPct).toFixed(1)}%;`}></span>
-        </span>
-        <span class="muted">${stats.lastPromptTokens.toLocaleString()} / ${(stats.contextCapTokens / 1000).toFixed(0)}K</span>
-      </span>
-      <span class="status-item">
-        <span class="status-label">cache</span>
-        <span class=${stats.cacheHitRatio >= 0.9 ? "status-ok" : stats.cacheHitRatio >= 0.6 ? "status-warn" : "status-err"}>
-          ${(stats.cacheHitRatio * 100).toFixed(1)}%
-        </span>
-      </span>
-      <span class="status-item">
-        <span class="status-label">turn</span>
-        <code>${fmtUsd(stats.lastTurnCostUsd)}</code>
-      </span>
-      <span class="status-item">
-        <span class="status-label">session</span>
-        <code>${fmtUsd(stats.totalCostUsd)}</code>
-        <span class="muted" style="font-size: 10px;">
-          (${stats.turns} turn${stats.turns === 1 ? "" : "s"})
-        </span>
-      </span>
-      ${
-        balance
-          ? html`
-          <span class="status-item">
-            <span class="status-label">balance</span>
-            <code>${balance.total_balance} ${balance.currency}</code>
-          </span>
-        `
-          : null
-      }
-    </div>
-  `;
-}
-
-// ---------- Sessions browser ----------
-
-function SessionsPanel() {
-  const { data, error, loading } = usePoll("/sessions", 5000);
-  const [open, setOpen] = useState(null); // { name, messages } or null
-  const [openLoading, setOpenLoading] = useState(false);
-
-  const view = useCallback(async (name) => {
-    setOpen({ name, messages: null });
-    setOpenLoading(true);
-    try {
-      const detail = await api(`/sessions/${encodeURIComponent(name)}`);
-      setOpen({ name, messages: detail.messages });
-    } catch (err) {
-      setOpen({ name, messages: null, error: err.message });
-    } finally {
-      setOpenLoading(false);
-    }
-  }, []);
-
-  if (loading && !data) return html`<div class="boot">loading sessions…</div>`;
-  if (error) return html`<div class="notice err">sessions failed: ${error.message}</div>`;
-  const sessions = data.sessions ?? [];
-
-  if (open) {
-    return html`
-      <div>
-        <div class="panel-header">
-          <h2 class="panel-title">Session</h2>
-          <span class="panel-subtitle">${open.name}</span>
-          <button onClick=${() => setOpen(null)} style="margin-left: auto;">← back</button>
-        </div>
-        ${
-          openLoading
-            ? html`<div class="boot">loading transcript…</div>`
-            : open.error
-              ? html`<div class="notice err">${open.error}</div>`
-              : open.messages && open.messages.length > 0
-                ? html`
-                <div class="chat-feed" style="max-height: calc(100vh - 180px); overflow-y: auto;">
-                  ${open.messages.map(
-                    (m, i) => html`
-                    <${ChatMessage}
-                      key=${i}
-                      msg=${{
-                        id: `r-${i}`,
-                        role:
-                          m.role === "tool"
-                            ? "tool"
-                            : m.role === "assistant"
-                              ? "assistant"
-                              : m.role === "user"
-                                ? "user"
-                                : "info",
-                        text: m.content ?? "",
-                        toolName: m.toolName,
-                      }}
-                      streaming=${false}
-                    />
-                  `,
-                  )}
-                </div>
-              `
-                : html`<div class="empty">empty transcript.</div>`
-        }
-      </div>
-    `;
-  }
-
-  return html`
-    <div>
-      <div class="panel-header">
-        <h2 class="panel-title">Sessions</h2>
-        <span class="panel-subtitle">${sessions.length} saved · click to read</span>
-      </div>
-      ${
-        sessions.length === 0
-          ? html`<div class="empty">No saved sessions yet.</div>`
-          : html`
-          <table>
-            <thead>
-              <tr>
-                <th>name</th>
-                <th class="numeric">messages</th>
-                <th class="numeric">size</th>
-                <th class="numeric">last touched</th>
-              </tr>
-            </thead>
-            <tbody>
-              ${sessions.map(
-                (s) => html`
-                <tr key=${s.name} onClick=${() => view(s.name)} style="cursor: pointer;">
-                  <td><code>${s.name}</code></td>
-                  <td class="numeric">${fmtNum(s.messageCount)}</td>
-                  <td class="numeric">${fmtBytes(s.size)}</td>
-                  <td class="numeric muted">${fmtRelativeTime(s.mtime)}</td>
-                </tr>
-              `,
-              )}
-            </tbody>
-          </table>
-        `
-      }
-    </div>
-  `;
-}
-
-// ---------- Plans archive ----------
-
-// ---------- shell ----------
-
 const TABS = [
-  {
-    id: "chat",
-    name: "Chat",
-    glyph: "◆",
-    panel: () => html`<${ChatPanel} />`,
-    ready: true,
-    badge: null,
-  },
-  {
-    id: "overview",
-    name: "Overview",
-    glyph: "◈",
-    panel: () => html`<${OverviewPanel} />`,
-    ready: true,
-    badge: null,
-  },
-  {
-    id: "usage",
-    name: "Usage",
-    glyph: "$",
-    panel: () => html`<${UsagePanel} />`,
-    ready: true,
-    badge: null,
-  },
-  {
-    id: "sessions",
-    name: "Sessions",
-    glyph: "›",
-    panel: () => html`<${SessionsPanel} />`,
-    ready: true,
-    badge: null,
-  },
-  {
-    id: "plans",
-    name: "Plans",
-    glyph: "P",
-    panel: () => html`<${PlansPanel} />`,
-    ready: true,
-    badge: null,
-  },
-  {
-    id: "tools",
-    name: "Tools",
-    glyph: "▣",
-    panel: () => html`<${ToolsPanel} />`,
-    ready: true,
-    badge: null,
-  },
-  {
-    id: "permissions",
-    name: "Permissions",
-    glyph: "▎",
-    panel: () => html`<${PermissionsPanel} />`,
-    ready: true,
-    badge: null,
-  },
-  {
-    id: "health",
-    name: "System",
-    glyph: "+",
-    panel: () => html`<${SystemPanel} />`,
-    ready: true,
-    badge: null,
-  },
-  {
-    id: "semantic",
-    name: "Semantic",
-    glyph: "≈",
-    panel: () => html`<${SemanticPanel} />`,
-    ready: true,
-    badge: null,
-  },
-  {
-    id: "mcp",
-    name: "MCP",
-    glyph: "M",
-    panel: () => html`<${McpPanel} />`,
-    ready: true,
-    badge: null,
-  },
-  {
-    id: "skills",
-    name: "Skills",
-    glyph: "S",
-    panel: () => html`<${SkillsPanel} />`,
-    ready: true,
-    badge: null,
-  },
-  {
-    id: "memory",
-    name: "Memory",
-    glyph: "·",
-    panel: () => html`<${MemoryPanel} />`,
-    ready: true,
-    badge: null,
-  },
-  {
-    id: "hooks",
-    name: "Hooks",
-    glyph: "H",
-    panel: () => html`<${HooksPanel} />`,
-    ready: true,
-    badge: null,
-  },
-  {
-    id: "settings",
-    name: "Settings",
-    glyph: "⌘",
-    panel: () => html`<${SettingsPanel} />`,
-    ready: true,
-    badge: null,
-  },
+  { id: "chat", name: "Chat", glyph: "◆", panel: () => html`<${ChatPanel} />`, ready: true, badge: null },
+  { id: "overview", name: "Overview", glyph: "◈", panel: () => html`<${OverviewPanel} />`, ready: true, badge: null },
+  { id: "usage", name: "Usage", glyph: "$", panel: () => html`<${UsagePanel} />`, ready: true, badge: null },
+  { id: "sessions", name: "Sessions", glyph: "›", panel: () => html`<${SessionsPanel} />`, ready: true, badge: null },
+  { id: "plans", name: "Plans", glyph: "P", panel: () => html`<${PlansPanel} />`, ready: true, badge: null },
+  { id: "tools", name: "Tools", glyph: "▣", panel: () => html`<${ToolsPanel} />`, ready: true, badge: null },
+  { id: "permissions", name: "Permissions", glyph: "▎", panel: () => html`<${PermissionsPanel} />`, ready: true, badge: null },
+  { id: "health", name: "System", glyph: "+", panel: () => html`<${SystemPanel} />`, ready: true, badge: null },
+  { id: "semantic", name: "Semantic", glyph: "≈", panel: () => html`<${SemanticPanel} />`, ready: true, badge: null },
+  { id: "mcp", name: "MCP", glyph: "M", panel: () => html`<${McpPanel} />`, ready: true, badge: null },
+  { id: "skills", name: "Skills", glyph: "S", panel: () => html`<${SkillsPanel} />`, ready: true, badge: null },
+  { id: "memory", name: "Memory", glyph: "·", panel: () => html`<${MemoryPanel} />`, ready: true, badge: null },
+  { id: "hooks", name: "Hooks", glyph: "H", panel: () => html`<${HooksPanel} />`, ready: true, badge: null },
+  { id: "settings", name: "Settings", glyph: "⌘", panel: () => html`<${SettingsPanel} />`, ready: true, badge: null },
 ];
 
 function App() {
   const [activeId, setActiveId] = useState("chat");
-  const [sidebarOpen, setSidebarOpen] = useState(false); // mobile drawer
-  // Desktop "icon only" collapse — narrow sidebar that shows just the
-  // glyphs. Persisted so the choice survives reload.
+  const [sidebarOpen, setSidebarOpen] = useState(false);
+  // Desktop "icon only" collapse — persisted so the choice survives reload.
   const [sidebarCollapsed, setSidebarCollapsed] = useState(() => {
     try {
       return localStorage.getItem("rx.sidebarCollapsed") === "1";
@@ -1575,8 +60,6 @@ function App() {
   }, [sidebarCollapsed]);
   const active = TABS.find((t) => t.id === activeId) ?? TABS[0];
 
-  // Esc anywhere closes the mobile drawer (modals already handle their
-  // own Esc). On desktop the drawer is always-open so this is a no-op.
   useEffect(() => {
     const onKey = (e) => {
       if (e.key === "Escape") setSidebarOpen(false);
@@ -1585,7 +68,6 @@ function App() {
     return () => window.removeEventListener("keydown", onKey);
   }, []);
 
-  // Cross-component navigation — sidebar-tab switching on `navigate-tab`.
   useEffect(() => {
     const onNav = (ev) => {
       const id = ev.detail?.tabId;
@@ -1597,7 +79,7 @@ function App() {
 
   const pickTab = useCallback((id) => {
     setActiveId(id);
-    setSidebarOpen(false); // collapse drawer after pick on mobile
+    setSidebarOpen(false);
   }, []);
 
   return html`
